+++ BoredGames.sln
@@ -1,12 +1,16 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.WebAPI", "src\BoredGames.WebAPI\BoredGames.WebAPI.csproj", "{C144B3C8-715D-4D4B-A942-051A065C303D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.Api", "src\BoredGames.Api\BoredGames.Api.csproj", "{C144B3C8-715D-4D4B-A942-051A065C303D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.Common", "src\BoredGames.Common\BoredGames.Common.csproj", "{8B338131-FCA2-4854-91C4-FCCF32F7A9D8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.Core", "src\BoredGames.Core\BoredGames.Core.csproj", "{8B338131-FCA2-4854-91C4-FCCF32F7A9D8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.Apologies", "src\BoredGames.Apologies\BoredGames.Apologies.csproj", "{564F57B0-21E8-48F6-88C6-B92D172F65B3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.Games.Apologies", "src\BoredGames.Games.Apologies\BoredGames.Games.Apologies.csproj", "{564F57B0-21E8-48F6-88C6-B92D172F65B3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.UnitTests.Apologies", "tests\unit\BoredGames.UnitTests.Apologies\BoredGames.UnitTests.Apologies.csproj", "{8C5FE7D9-ECA9-4CC8-A3AB-1AC87CCEDF62}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Games", "Games", "{215E13D5-F254-4CFF-AAB7-D708F2EE7326}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BoredGames.UnitTests.Apologies", "test\unit\BoredGames.UnitTests.Apologies\BoredGames.UnitTests.Apologies.csproj", "{BE256B40-7B62-4AFF-850C-1F76E3BDE8DC}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{F5540FA6-7FEE-4409-A569-DDB90134837A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -26,11 +30,13 @@ Global
 		{564F57B0-21E8-48F6-88C6-B92D172F65B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{564F57B0-21E8-48F6-88C6-B92D172F65B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{564F57B0-21E8-48F6-88C6-B92D172F65B3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8C5FE7D9-ECA9-4CC8-A3AB-1AC87CCEDF62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8C5FE7D9-ECA9-4CC8-A3AB-1AC87CCEDF62}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8C5FE7D9-ECA9-4CC8-A3AB-1AC87CCEDF62}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8C5FE7D9-ECA9-4CC8-A3AB-1AC87CCEDF62}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BE256B40-7B62-4AFF-850C-1F76E3BDE8DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BE256B40-7B62-4AFF-850C-1F76E3BDE8DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BE256B40-7B62-4AFF-850C-1F76E3BDE8DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BE256B40-7B62-4AFF-850C-1F76E3BDE8DC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{564F57B0-21E8-48F6-88C6-B92D172F65B3} = {215E13D5-F254-4CFF-AAB7-D708F2EE7326}
+		{BE256B40-7B62-4AFF-850C-1F76E3BDE8DC} = {F5540FA6-7FEE-4409-A569-DDB90134837A}
 	EndGlobalSection
 EndGlobal
