--- misc/API Endpoints/BoredGames.http
@@ -1,6 +1,0 @@
-@BoredGames_HostAddress = http://localhost:5251
-
-GET {{BoredGames_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
