--- misc/MockData/Example11Movelist.json
@@ -1,21 +1,0 @@
-{
-  "card": 11,
-  "moveset" : [
-    {
-      "piece" : "d_3",
-      "moves" : [
-        "c_7",
-        "c_13",
-        "d_14"
-      ]
-    },
-    {
-      "piece" : "a_9",
-      "moves" : [
-        "c_7",
-        "c_13",
-        "a_s3"
-      ]
-    }
-  ]
-}
