--- misc/MockData/ExampleMovelist.json
@@ -1,17 +1,0 @@
-{
-  "card": 2,
-  "moveset" : [
-    {
-      "piece" : "a_S",
-      "moves" : [
-        "a_4"
-      ]
-    },
-    {
-      "piece" : "c_10",
-      "moves" : [
-        "c_12"
-      ]
-    }
-  ]
-}
