--- misc/MockData/ExampleSorryMovelist.json
@@ -1,12 +1,0 @@
-{
-  "card": 0,
-  "moveset" : [
-    {
-      "piece" : "a_S",
-      "moves" : [
-        "c_7",
-        "c_13"
-      ]
-    }
-  ]
-}
