+++ src/BoredGames.Core/BoredGames.Core.csproj
