+++ src/BoredGames.Core/Game/IGameConfig.cs
@@ -1,0 +1,12 @@
+using System.Collections.Immutable;
+
+namespace BoredGames.Core.Game;
+
+public interface IGameConfig
+{
+    public GameTypes GameType { get; }
+    public int MinPlayerCount { get; }
+    public int MaxPlayerCount { get; }
+    
+    public GameBase CreateGameInstance(ImmutableList<Player> players);
+}
