+++ src/BoredGames.Games.Apologies/Models/ActionResponses.cs
@@ -1,0 +1,18 @@
+using BoredGames.Core.Game;
+using JetBrains.Annotations;
+
+namespace BoredGames.Games.Apologies.Models;
+
+public static class ActionResponses
+{
+    [UsedImplicitly]
+    public record DrawCardResponse(
+        int CardDrawn,
+        IEnumerable<GenericComponents.Moveset> Movesets) : IGameActionResponse;
+
+    [UsedImplicitly]
+    public record EndgameStatsResponse(
+        IEnumerable<int> MovesMade,
+        IEnumerable<int> PawnsKilled, 
+        long GameTimeElapsed) : IGameActionResponse;
+}
