--- src/BoredGames.WebAPI/BoredGames.WebAPI.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net9.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-        <RootNamespace>BoredGames</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.6" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="9.0.3" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\BoredGames.Apologies\BoredGames.Apologies.csproj" />
-        <ProjectReference Include="..\BoredGames.Common\BoredGames.Common.csproj" />
-    </ItemGroup>
-
-
-</Project>
