--- src/BoredGames.WebAPI/Models/RoomSnapshot.cs
@@ -1,8 +1,0 @@
-using BoredGames.Common.Game;
-using BoredGames.Common.Room;
-using JetBrains.Annotations;
-
-namespace BoredGames.Models;
-
-[UsedImplicitly]
-public record RoomSnapshot(int ViewNum, GameRoom.State State, IEnumerable<string> Players, IGameSnapshot? GameSnapshot);
